+++ src/main/java/com/example/comprehensivedegisn/adapter/repository/apart/QuerydslApartmentTransactionRepository.java
@@ -5,10 +5,10 @@ import com.example.comprehensivedegisn.adapter.domain.Gu;
 import com.example.comprehensivedegisn.adapter.domain.PredictStatus;
 import com.example.comprehensivedegisn.adapter.order.CustomPageable;
 import com.example.comprehensivedegisn.adapter.order.CustomPageImpl;
-import com.example.comprehensivedegisn.dto.SearchApartNameResponse;
-import com.example.comprehensivedegisn.dto.SearchAreaResponse;
-import com.example.comprehensivedegisn.dto.SearchResponseRecord;
-import com.example.comprehensivedegisn.dto.SearchCondition;
+import com.example.comprehensivedegisn.dto.response.SearchApartNameResponse;
+import com.example.comprehensivedegisn.dto.response.SearchAreaResponse;
+import com.example.comprehensivedegisn.dto.response.SearchResponseRecord;
+import com.example.comprehensivedegisn.dto.request.SearchCondition;
 import com.querydsl.core.types.Projections;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQuery;
