+++ src/main/java/com/example/comprehensivedegisn/dto/request/SearchApartNameRequest.java
@@ -1,4 +1,4 @@
-package com.example.comprehensivedegisn.dto;
+package com.example.comprehensivedegisn.dto.request;
 
 import com.example.comprehensivedegisn.adapter.domain.Gu;
 import lombok.AllArgsConstructor;
@@ -9,22 +9,13 @@ import org.springframework.util.StringUtils;
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
-public class SearchAreaRequest{
+public class SearchApartNameRequest {
 
     private Gu gu;
     private String dong;
-    private String apartmentName;
 
     public boolean isNotValid() {
-        return isGuNotValid() | isDongNotValidate() || isApartmentNameNotValid();
-    }
-
-    private boolean isApartmentNameNotValid() {
-        return !StringUtils.hasText(apartmentName);
-    }
-
-    private boolean isDongNotValidate() {
-        return !StringUtils.hasText(dong);
+        return isGuNotValid() | !StringUtils.hasText(dong);
     }
 
     private boolean isGuNotValid() {
