+++ src/main/java/com/example/comprehensivedegisn/dto/response/SearchApartNameResponse.java
@@ -1,4 +1,4 @@
-package com.example.comprehensivedegisn.dto;
+package com.example.comprehensivedegisn.dto.response;
 
 public record SearchApartNameResponse(String apartmentName) {
 }
