+++ src/main/java/com/example/comprehensivedegisn/dto/response/SearchAreaResponse.java
@@ -1,4 +1,4 @@
-package com.example.comprehensivedegisn.dto;
+package com.example.comprehensivedegisn.dto.response;
 
 public record SearchAreaResponse(double areaForExclusiveUse) {
 }
